renaming models to upperCaseFirst, but parcel ...
* parcel now fails. It shows errors like
  `SyntaxError: Unexpected token }` in built js files.
  Rebuilding leads to other syntax error. I really want to replace soon
  parcel. It has so weired bugs.
