+++ examples/src/triangles-web-with-d3.ts
@@ -2,7 +2,7 @@ import * as d3 from 'd3';
 import Delaunator from 'delaunator';
 
 import { PointsGrid } from '../../src';
-import { Point } from '../../src/model/types';
+import { Point } from '../../src/model';
 
 interface TrianglesWebOptions {
     svgSelector: string;
