+++ src/model/Point.ts
