+++ src/model/renderer.ts
@@ -1,4 +1,4 @@
-import { Renderable } from "./renderable";
+import { Renderable } from "./Renderable";
 
 /** What a common Renderer needs to have defined. Idea: to add later CanvasRenderer next to SvgRenderer which works the same way.
  * The developer can just switch the renderer if desired (e.g. to test if performance is better with other renderer)
